feat: se cambian de lugar los enumerables, no todos, los que se podía, ahora están en la carpeta que les corresponde y también se modificó su estructura permitiendo una implementación más sencilla
