+++ src/main/java/com/healink/integrador/domain/atenciones_medicas/AtencionMedica.java
@@ -10,7 +10,6 @@ import java.sql.Timestamp;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.healink.integrador.core.entity.EntidadAuditable;
-import com.healink.integrador.enums.EstadoAtencionMedica;
 
 
 @Entity
