+++ src/main/java/com/healink/integrador/domain/atenciones_medicas/EstadoAtencionMedica.java
@@ -1,0 +1,7 @@
+package com.healink.integrador.domain.atenciones_medicas;
+
+public enum EstadoAtencionMedica {
+    EN_PROCESO,
+    COMPLETADA,
+    CANCELADA
+}
