+++ src/main/java/com/healink/integrador/domain/diagnosticos/SeveridadDiagnostico.java
@@ -1,0 +1,7 @@
+package com.healink.integrador.domain.diagnosticos;
+
+public enum SeveridadDiagnostico {
+    LEVE, 
+    MODERADAADA,
+    GRAVE
+}
