+++ src/main/java/com/healink/integrador/domain/prescripciones/Prescripcion.java
@@ -8,7 +8,6 @@ import lombok.NoArgsConstructor;
 
 import com.healink.integrador.core.entity.EntidadAuditable;
 import com.healink.integrador.domain.diagnosticos.Diagnostico;
-import com.healink.integrador.enums.TipoPrescripcion;
 
 import java.time.LocalDate;
