+++ src/main/java/com/healink/integrador/domain/prescripciones/PrescripcionDTO.java
@@ -3,7 +3,6 @@ package com.healink.integrador.domain.prescripciones;
 import java.time.LocalDate;
 
 import com.healink.integrador.core.dto.DTOBase;
-import com.healink.integrador.enums.TipoPrescripcion;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
