+++ src/main/java/com/healink/integrador/domain/prescripciones/TipoPrescripcion.java
@@ -1,0 +1,8 @@
+package com.healink.integrador.domain.prescripciones;
+
+public enum TipoPrescripcion {
+    MEDICAMENTO,
+    ESTILO_VIDA,
+    ACTIVIDAD_FISICA,
+    DIETA;
+}
