+++ src/main/java/com/healink/integrador/domain/recomendaciones/NivelPrioridad.java
@@ -1,0 +1,7 @@
+package com.healink.integrador.domain.recomendaciones;
+
+public enum NivelPrioridad {
+    ALTA,
+    MEDIA,
+    BAJA
+}
