+++ src/main/java/com/healink/integrador/domain/recomendaciones/Recomendacion.java
@@ -8,8 +8,6 @@ import lombok.NoArgsConstructor;
 
 import com.healink.integrador.core.entity.EntidadAuditable;
 import com.healink.integrador.domain.diagnosticos.Diagnostico;
-import com.healink.integrador.enums.NivelPrioridad;
-import com.healink.integrador.enums.TipoRecomendaciones;
 
 @Entity
 @Table(name = "RECOMENDACIONES")
