+++ src/main/java/com/healink/integrador/domain/recomendaciones/TipoRecomendaciones.java
@@ -1,0 +1,7 @@
+package com.healink.integrador.domain.recomendaciones;
+
+public enum TipoRecomendaciones {
+    MEDICAMENTO,
+    ESTILO_VIDA,
+    PREVENCION
+}
