+++ src/main/java/com/healink/integrador/domain/seguimientos/EstadoSeguimiento.java
@@ -1,0 +1,7 @@
+package com.healink.integrador.domain.seguimientos;
+
+public enum EstadoSeguimiento {
+    PENDIENTE,
+    REALIZADO,
+    CANCELADO
+}
