+++ src/main/java/com/healink/integrador/domain/seguimientos/NivelPrioridad.java
@@ -1,0 +1,7 @@
+package com.healink.integrador.domain.seguimientos;
+
+public enum NivelPrioridad {
+    ALTA,
+    MEDIA,
+    BAJA
+}
