+++ src/main/java/com/healink/integrador/domain/seguimientos/Seguimiento.java
@@ -8,9 +8,7 @@ import lombok.NoArgsConstructor;
 
 import com.healink.integrador.core.entity.EntidadAuditable;
 import com.healink.integrador.domain.atenciones_medicas.AtencionMedica;
-import com.healink.integrador.enums.EstadoSeguimiento;
-import com.healink.integrador.enums.NivelPrioridad;
-import com.healink.integrador.enums.TipoSeguimiento;
+import com.healink.integrador.domain.recomendaciones.NivelPrioridad;
 
 import java.time.LocalDate;
