+++ src/main/java/com/healink/integrador/domain/seguimientos/TipoSeguimiento.java
@@ -1,0 +1,7 @@
+package com.healink.integrador.domain.seguimientos;
+
+public enum TipoSeguimiento {
+    LLAMADA,
+    SMS,
+    PRESENCIAL;
+}
