+++ src/main/java/com/healink/integrador/enums/TipoIdentificacion.java
@@ -1,32 +1,8 @@
 package com.healink.integrador.enums;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-
 public enum TipoIdentificacion {
-    CC("CC"),
-    TI("TI"),
-    NIT("NIT"),
-    RCN("RCN");
-
-    private final String value;
-
-    TipoIdentificacion(String value) {
-        this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-        return value;
-    }
-
-    @JsonCreator
-    public static TipoIdentificacion fromValue(String value) {
-        for (TipoIdentificacion tipo : TipoIdentificacion.values()) {
-            if (tipo.value.equalsIgnoreCase(value)) {
-                return tipo;
-            }
-        }
-        throw new IllegalArgumentException("Tipo de identificación inválido: " + value);
-    }
+    CC,
+    TI,
+    NIT,
+    RCN
 }
